+++ src/Fiv/Form/Validator/ValidatorInterface.php
@@ -1,0 +1,22 @@
+<?php
+  namespace Fiv\Form\Validator;
+
+
+  /**
+   *
+   * @package Fiv\Form\Validator
+   */
+  interface ValidatorInterface {
+
+    /**
+     * @return array
+     */
+    public function getErrors();
+
+
+    /**
+     * @param array|string $value
+     */
+    public function isValid($value);
+
+  }