+++ CHANGELOG.md
@@ -7,5 +7,6 @@
 
 ### Deprecated
 - \Fiv\Form\Validator\Base::i
+- \Fiv\Form\Validator\Base use \Fiv\Form\Validator\BaseValidator
 - \Fiv\Form\Filter\Base::i
 - \Fiv\Form\Filter use \Fiv\Form\Filter\FilterInterface instead